refactor(security): implementar UserResponseDTO y agregar validaciones Jakarta a los DTOs
Descripción adicional para el cuerpo del commit (opcional):

Cambia los controladores UserController y AuthController para retornar UserResponseDTO en lugar de la entidad User de JPA, previniendo la fuga de contraseñas u otros datos internos.
Agrega anotaciones de validación de Jakarta con mensajes personalizados en los DTOs de entrada.
Adapta las pruebas unitarias e de integración al nuevo flujo basado en DTOs.
Ajusta la configuración de Bucket4j y el script de semillas de base de datos (seed_testing.sql) para pruebas unitarias.

+++ backend/src/main/java/com/g9latam/team62/fintech_api/model/FinancialProfileHistory.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 import jakarta.persistence.Entity;
@@ -28,6 +29,6 @@ public class FinancialProfileHistory {
     
     @Enumerated(EnumType.STRING)
     private FinancialProfile financialProfile;
-    private Double profileAccuracy;
+    private BigDecimal profileAccuracy;
     private LocalDateTime createdAt = LocalDateTime.now();
 }
